Refactor: 로그아웃 링크 버튼 코드 및 디자인 리팩토링
