+++ src/components/common/index.ts
@@ -4,3 +4,4 @@ export { default as FixedBottomButton } from './FixedBottomButton';
 export { default as Carousel } from './Carousel';
 export { default as Header } from './Header';
 export { default as Modal } from './Modal';
+export { default as LogoutBtn } from './LogoutBtn';
